requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
cryptography==36.0.1
defusedxml==0.7.1
dj-static==0.0.6
Django==4.0.1
django-appconf==1.0.5
django-cleanup==6.0.0
django-cors-headers==3.11.0
django-filter==21.1
django-rest-auth==0.9.5
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.10.1
Markdown==3.3.6
multidict==5.2.0
oauthlib==3.1.1
Pillow==9.0.0
pycparser==2.21
Pygments==2.11.2
PyJWT==2.3.0
PySocks==1.7.1
python-dotenv==0.19.2
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
sentry-sdk==1.5.4
six==1.16.0
sqlparse==0.4.2
static3==0.7.0
urllib3==1.26.8
whitenoise==5.3.0
zipp==3.7.0
